+++ pkg/alertmanager/state_replication_test.go
@@ -143,7 +143,7 @@ func TestStateReplication(t *testing.T) {
 				require.NoError(t, s.WaitReady(ctx))
 			}
 
-			ch := s.AddState("nflog", &fakeState{}, reg)
+			ch := s.AddState("nflog:user-1", &fakeState{}, reg)
 
 			part := tt.message
 			d, err := part.Marshal()
@@ -166,10 +166,10 @@ alertmanager_state_fetch_replica_state_failed_total 0
 alertmanager_state_fetch_replica_state_total 1
 # HELP alertmanager_partial_state_merges_failed_total Number of times we have failed to merge a partial state received for a key.
 # TYPE alertmanager_partial_state_merges_failed_total counter
-alertmanager_partial_state_merges_failed_total{key="nflog"} 0
+alertmanager_partial_state_merges_failed_total{type="nflog"} 0
 # HELP alertmanager_partial_state_merges_total Number of times we have received a partial state to merge for a key.
 # TYPE alertmanager_partial_state_merges_total counter
-alertmanager_partial_state_merges_total{key="nflog"} 0
+alertmanager_partial_state_merges_total{type="nflog"} 0
 # HELP alertmanager_state_initial_sync_completed_total Number of times we have completed syncing initial state for each possible outcome.
 # TYPE alertmanager_state_initial_sync_completed_total counter
 alertmanager_state_initial_sync_completed_total{outcome="failed"} 0
@@ -181,10 +181,10 @@ alertmanager_state_initial_sync_completed_total{outcome="user-not-found"} 0
 alertmanager_state_initial_sync_total 1
 # HELP alertmanager_state_replication_failed_total Number of times we have failed to replicate a state to other alertmanagers.
 # TYPE alertmanager_state_replication_failed_total counter
-alertmanager_state_replication_failed_total{key="nflog"} 0
+alertmanager_state_replication_failed_total{type="nflog"} 0
 # HELP alertmanager_state_replication_total Number of times we have tried to replicate a state to other alertmanagers.
 # TYPE alertmanager_state_replication_total counter
-alertmanager_state_replication_total{key="nflog"} 1
+alertmanager_state_replication_total{type="nflog"} 1
 	`),
 					"alertmanager_state_fetch_replica_state_failed_total",
 					"alertmanager_state_fetch_replica_state_total",